Add cast to fix build error

+++ source/blender/io/usd/intern/usd_reader_mesh.cc
@@ -541,7 +541,7 @@ void USDMeshReader::process_normals_vertex_varying(Mesh *mesh)
 
   BLI_STATIC_ASSERT(sizeof(normals_[0]) == sizeof(float3), "Expected float3 normals size");
   bke::mesh_set_custom_normals_from_verts(
-      *mesh, {reinterpret_cast<float3 *>(normals_.data()), normals_.size()});
+      *mesh, {reinterpret_cast<float3 *>(normals_.data()), int64_t(normals_.size())});
 }
 
 void USDMeshReader::process_normals_face_varying(Mesh *mesh) const
